+++ libbeat/processors/translate_ldap_attribute/docs/translate_ldap_attribute.asciidoc
@@ -2,29 +2,32 @@
 === Translate GUID
 
 ++++
-<titleabbrev>translate_guid</titleabbrev>
+<titleabbrev>translate_ldap_attribute</titleabbrev>
 ++++
 
-The `translate_guid` processor translates an LDAP Global Unique Identifier (GUID)
-into its common name.
+The `translate_ldap_attribute` processor translates an LDAP attributes between eachother.
+It is typically used to translate AD Global Unique Identifiers (GUID)
+into their common names.
 
 Every object on an Active Directory or an LDAP server is issued a GUID. Internal processes
 refer to their GUID's rather than the object's name and these values
 sometimes appear in logs.
 
-If the GUID is invalid (malformed) or does not map to any object on the domain
+If the search attribute is invalid (malformed) or does not map to any object on the domain
 then this will result in the processor returning an error unless `ignore_failure`
 is set.
-
-Both the attribute referring to the GUID and the one to map it to, can be changed.
 
 The result of this operation is an array of values, given that a single attribute
 can hold multiple values.
 
+Note: the search attribute is expected to map to a single object. If it doesn't,
+no error will be returned, but only results of the first entry will be added
+to the event.
+
 [source,yaml]
 ----
 processors:
-  - translate_guid:
+  - translate_ldap_attribute:
       field: winlog.event_data.ObjectGuid
       ldap_address: "ldap://"
       ldap_base_dn: "dc=example,dc=com"
@@ -32,7 +35,7 @@ processors:
       ignore_failure: true
 ----
 
-The `translate_guid` processor has the following configuration settings:
+The `translate_ldap_attribute` processor has the following configuration settings:
 
 .Translate GUID options
 [options="header"]
@@ -44,8 +47,8 @@ The `translate_guid` processor has the following configuration settings:
 | `ldap_base_dn`           | yes      |              | LDAP base DN. eg: `dc=example,dc=com`
 | `ldap_bind_user`         | no       |              | LDAP user.
 | `ldap_bind_password`     | no       |              | LDAP password.
-| `ldap_guid_attribute`    | yes      | `objectGUID` | LDAP attribute containing the GUID.
-| `ldap_mapped_attribute`  | yes      | `cn`         | LDAP attribute that will be mapped to the GUID.
+| `ldap_search_attribute`  | yes      | `objectGUID` | LDAP attribute to search by.
+| `ldap_mapped_attribute`  | yes      | `cn`         | LDAP attribute to map to.
 | `ldap_search_time_limit` | no       | 30           | LDAP search time limit in seconds.
 | `ldap_ssl`*              | no       | 30           | LDAP TLS/SSL connection settings.
 | `ignore_missing`         | no       | false        | Ignore errors when the source field is missing.
@@ -54,7 +57,8 @@ The `translate_guid` processor has the following configuration settings:
 
 &#42; Also see <<configuration-ssl>> for a full description of the `ldap_ssl` options.
 
-If the searches are slow or you expect a high amount of GUID's to be found, consider using a cache processor to speed processing:
+If the searches are slow or you expect a high amount of different key attributes to be found,
+consider using a cache processor to speed processing:
 
 
 [source,yaml]
@@ -72,7 +76,7 @@ processors:
       not:
         - has_fields: winlog.common_name
     then:
-      - translate_guid:
+      - translate_ldap_attribute:
           field: winlog.event_data.ObjectGuid
           target_field: winlog.common_name
           ldap_address: "ldap://"
